+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Chores
 
+- Refactored model building and validation to work on internal model elements
+  instead
 - Refactored AST and internal model to store order dependent child nodes in
   separate fields/properties to allow for easier runtime modifications without
   having to maintain implicit ordering